+++ wmfdata/spark.py
@@ -15,7 +15,13 @@ from wmfdata.utils import (
 """
 Will be used for findspark.init().
 """
-SPARK_HOME = os.environ.get("SPARK_HOME", "/usr/lib/spark2")
+SPARK_HOME = os.environ.get("SPARK_HOME", "/usr/lib/spark3")
+if SPARK_HOME == "/usr/lib/spark2":
+    warnings.warn(
+        "Spark2 has been deprecated. Please upgrade your jobs to Spark3. "
+        "See https://phabricator.wikimedia.org/T318367 for details.",
+        category=FutureWarning
+    )
 
 """
 Predefined spark sessions and configs for use with the get_session and run functions.